Succeeded the queue implementation.

+++ Main/queue.h
@@ -10,7 +10,7 @@
 typedef struct linklist_t
 {
     void *data;
-    uint8_t len;
+    size_t len;
     struct linklist_t *next;
 } linklist_t;
 
@@ -18,14 +18,14 @@ typedef struct
 {
     linklist_t *head;  // First mover on new item.
     linklist_t *tail;  // Point to the first data in queue.
-    uint8_t len;
+    size_t len;
 } queue_t;
 
 extern void q_init(queue_t *q);
-extern linklist_t *q_enqueue(queue_t *q, void *data, uint8_t len);
-extern uint8_t q_dequeue(queue_t *q, void *data, uint8_t maxlen);
-extern uint8_t q_length(queue_t *q);
-extern linklist_t *q_item(queue_t *q, uint8_t index);
+extern linklist_t  *q_enqueue(queue_t *q, void *data, size_t len);
+extern size_t       q_dequeue(queue_t *q, void *data, size_t maxlen);
+extern size_t       q_length(queue_t *q);
+extern linklist_t  *q_item(queue_t *q, uint8_t index);
 
 
 #endif  // __QUEUE_H__
